api/requirements.txt
# This file is auto-generated by tools/extract-requirements.sh.
# Please do not edit this file, edit the setup.py file!
coreapi
django~=3.0
django-filter
djangorestframework
drf-spectacular
markdown
psycopg2
